Move row/section size calculation to a function
Since we are now loading asynchronously, Repeater ends up reporting a
different item count from the actual instantiated items. This means we
need to deal with itemAt returning null, which means bindings are not
created and updates do not happen. So instead use a funciton which
calculates all this and trigger that in a bunch of places.