Pop pages that were opened during editing after editing
If a face configuration page was opened during editing and not closed when
exiting editing, it will break when we switch back to pageContents with the
effect that changed face settings are not applied anymore until the app is
restarted. This manifests in error messages such as
FaceConfigurationPage.qml:189: TypeError: Cannot read property 'controller' of null
When we properly pop the page of the stack it works as expected.