Allow the user to display percentage divided by number of cores
Introduces two new display styles that are shown if the Unit of an attribute is
percent and the maximum is more than 100%. These will simply divide the displayed
value by the maximum. For example if on a system with four cpu cores the usage of
one process is 125%, it will be shown as 31.25%. The line chart maximum and
current value does not need to be changed because the visual represenation
of 125 of 400 and 31.25 of 100 is the same.