Properly handle empty groups when loading pages
Apparently KConfig doesn't have a way to mark a group as deleted, so if
we remove a row from a system-installed page, it will just mark all
entries in that group as deleted, but leave the group. Now we handle
that case and only add items for groups that contain something.