Add feature for exporting and adding a chart to the desktop
If a user is interested in some values that is not captured on the default
pages, they can set up a custom chart for them. If they want now to have the
same information/configuration as a widget on their desktop, they have to do
the same process twice. This adds an action so a chart can easily be exported
to the desktop if plasma is currently running. The action can be accessed from
the face configuration or by right clicking on a chart.