Avoid duplicate pages
If we know there is a page with the same name already, append an increasing
number until the name is unique. For the kns case if it were to download a
duplicate file, the kns dialog will prompt the user if they want to overwrite
the existing one. So we need to just reload the page which is now backed by
the downloaded file.