Workaround KDescendantsProxyModel not forwarding table headers properly
KDescendantsProxyModel does not forward table headers correctly so we
end up with empty headers. This has been fixed in the meantime but the
fix will only be in Frameworks 5.76. Since we want to be able to run
against Plasma 5.20 this works around the issue by not retrieving table
headers for the descendants model but from the sort model below it.