Fix filter invalidation
When we have matched columns we change what filterAcceptsRow passes.
This means we need to trigger re-evaluation of filterAcceptsRow not a layout change. A
layout change implies movement, but not a change in which rows exist.

Without this KDescendantsProxyModel gets out of sync as it correctly doesn't
refresh indexes on a layout change.

BUG: 428662