Fix possible crash when resetting a page
The loader is created by QML as is the data object through the PageModel and
the lifetimes of them are not fxed against one another.
BUG:428656
The backtrace I got and the place is different than in the bug report but I
found no further problems when resetting the overview page after changing a
sensor color while running with asan and ubsan.