Do not allow configuring the name column
We always want it at the start since the tree decoration assumes that there is
the icon at the start and also doesn't make sense to have it somewhere else.
This allows specifying a number of columns that cannot be configured and will
always be at the start of the table.