Display existing search text in search field for application/process pages
If the actions of the current page include the table actions, they will
be recreated when switching pages. However, the actual search string
will persist and still filter the table. So ensure we display the
existing filter string when we switch back to avoid confusion of why not
everything is shown.

BUG: 433324