Avoid showing "Processes: 0" in applications sidebar whilst loading
Showing "Processes: 0" whilst loading as it implies false information.

We should never have a cgroup with no processes, so it's safe to use
that as a way to query if we're loading.

The use of space instead of "" is there to avoid getting a runtime
warning from i18n about a missing argument.

A better fix will come when the underlying model is shared, and
therefore we will have all information immediately.

BUG: 433326


(cherry picked from commit 64d07214c15a08733f54d6e63adcadb4ffb7a274)