Fix scrollbar overlapping configuration page
The preferredWidth was to big because the scrollbar adds additional
padding in the internal hierachy of the page.
Also up the columnWidth otherwise we do not have enough space for
the wide mode of the formlayout inside the faceConfigUi.
BUG:434461
FIXED-IN:5.21.4

(cherry picked from commit c53c1c7b139b47d20ba347ca01f90b217dbf04ae)