Update ApplicationInformation immediately
When indexChanged was triggered and we went to update(), the valid
property was apparently not yet evaluated and so we would return and
do nothing. The data was only updated on the next dataChanged of the
model which could take any time up to one second.
We can just remove the property and use the condition directly. It
was not used anywhere else.