Avoid caching QModelIndexes
QModelIndexes can go out of date with no change notifications.

What we were seeing in the applications side bar is that the filter of
PIDs would change which effectively removes and replaces every delegate.

Delegates get re-used, if one is re-used at the same row and column the
modelIndex is not updated, however the internalId is completely wrong.
Having valid numbers and parents and wrong internalIds is an absolute
recipe for disaster!