Show child processes if we match the parent
In the process tree also show children if the we match the parent.
This is somewhat the opposite of recursive filtering ("show parent if a
child is shown"). Example:
 A
 |-B
 | |-D
 | |-E
 |-C
Currently if you search for "E", A, B and E will be shown, this is not
changed by this commit and the effect of the recursive filtering.
But if the filter is "B", only A and B will be shown.
With this commit A, B, D and E are shown.