Fix initial sort order
BaseTableView is duplicating code in TableViewHeader, this is non-ideal.

The other issue, is that we run the lambda whilst doing the initial
Repeater population. model.row is outdated when the lambda does run.

We're best off running the existing sort method on the delegate which
will keep everything in sync.

BUG: 437449