Push NewStuff pages to layers, not the stack
The NewStuff.Action component will simply call the push function on
whatever object it is given to work with as a page stack, and since the
Kirigami pagestack layers also has one such, we can just switch to that.

nb: The functionality is explicitly designed to work this way, to allow for
things like the kcm pagestack to work as well as the more featureful
Kirigami one.

BUG:437961