Clear layers when navigating
If there's a page open in a layer, navigating to a different system
monitor page has no effect since the layer hides the page change. So if
we have pages in a layer and the user clicks on a page to navigate to
it, clear the layers to close those pages and show the navigation
change.


(cherry picked from commit 53bf26623fb0f839a846559d94df2a563c1ae62f)