Allow changing row heights to one of three different modes
This introduces a "heightMode" property for page rows, which allows a
user to select one of three different behaviours for the height of a
row. Minimum will use the minimum size of the row, balanced matches the
old behaviour where we try to balance the row height between different
rows and finally maximum which will try to use as much space as
possible.

This allows for some different layouts where there's a single row of
text only sensors, most of the space is used for a process table or
similar constructions.

BUG: 441489
CCBUG: 441027