ApplicationsTable: Remove unneeded ProcessSortFilterModel
ProcessSortFilterModel handles filtering pids and to which user a process
belongs over a normal QSortFilterProxyModel. In doing so it also
sets the filterKeyColumn to nameColumn in findColumns(). Note that the name
of the name columns of the process and application table is different
("name" and "appName"), so the filterKeyColumn is also set by ApplicationTableView.
However this is racey depending on when the binding is evaluated.
Since we don't need the extra features of the ProcessSortFilterModel we can use
the existing SortFilterProxyModel in charge of filtering columns (something that
the ProcessSortFilterModel also did) to also handle the row filtering.
BUG:445544
FIXED-IN:5.23.4