Use the new dialog for replacing missing sensors
When a page is missing sensors, it may end up not displaying correctly.
This can happen because we have certain sensors that are bound to
physical hardware, which disappear or do not exist when we use a page
from the store.

Rather than trying to do a difficult (probably impossible) mapping
between sensors or trying to prevent uploading pages with
hardware-dependent sensors, provide a UI so this situation can be
resolved by the user.