Add a page option to allow preloading a page
For certain pages, it is useful to be able to load the page as soon as
possible so graphs can already start collecting data. This adds an
option to the page configuration that specifies when to load a page -
either on demand on during startup. By default pages will still be
loaded on demand.

BUG: 440402