table: Workaround a memory leak in QQmlPropertyCache
Apparently, the combination of a Connections object along with the
QAbstractItemModel::layoutChanged() signal leads to a memory leak inside
QQmlPropertyCache::createArgumentsObject(). To avoid that, put a proxy
signal in between that Connections can connect to.

The bug no longer exists in Qt 6 so this should be reverted after
switching to Qt 6.

BUG: 460370


(cherry picked from commit ebd169763aea87fa6203dc8f12ced639428525f4)