Provide a default for enabled columns
Due to a race in item instantiation if the process table view
is instantiated before the column configuration dialog the enabled
attributes will be initially empty. However we still enable columns
which then will be filtered out, confusing the proxy model stack
as there will be no valid parents anymore because the first column
has been filtered out but still claiming to have rows.
Delaying assignment of attributes causes a stack overflow due
to QSortFilterProxyModelPrivate::recursiveChildAcceptsRows calling
itself.
Providing here a default enabled column is a straightfoward fix
instead of trying to find the exact point of failure in the the proxymodels
provided by us and upstream and matches the default for columnDisplay.
BUG:470784