Workaround QTBUG-84858
In certain cases destroying a table cell will end up calling a virtual
method incorrectly. This was partially fixed upstream but only when the
Accessible attached object exists. To workaround the upstream bug,
ensure we create an instance of the attached object. As a bonus, this
also means we now set the right accessible role for table cells.

BUG: 464893