table: Fix sorting of tables using the header
A recent change in Qt enforced that the HorizontalHeaderView's
selectionModel always creates a header proxy model if an
abstractitemmodel is used for the HorizontalHeaderView's model. This
apparently broke setting the current column when using our own header
proxy model. However, we actually do not need the current column and
selection model handling, as it is just an integer property to keep
track of the column. So instead drop the selection model and track sort
column manually.

BUG: 479326


(cherry picked from commit e3ef0d767044930615334b10640a45165638eb26)