EditablePage: position loading placeholder overlay correctly
It was using negative margins to counteract the ones added to the page,
but was then re-parented to something that wasn't the page. If those two
things didn't have the exact same geometry, the overlay would be
mis-positioned.

Fix this by dropping the re-parenting, and instead manually set the Z to
something very high to ensure it always covers the content.

BUG: 488044
FIXED-IN: 6.1.1


(cherry picked from commit 3951c374c25fed5ef518e044273b47c6c5923e3d)

Co-authored-by: Nate Graham <nate@kde.org>