table: Use a separate synced TableView for row backgrounds
Rather than having each cell of the main TableView provide their own
background, this adds a separate TableView below the main view that is
synced to the main view and used to render the row backgrounds. This
ensures that we don't have backgrounds for individual cells but for full
rows, which prevents issues where row selection/hover is rendered
scattered over multiple cells due to cell reuse.