Move "Add New Page" UI into hamburger menu
It's not typical to allow creating a new item in a list from a fake "add
new" item in that list. This surfaces the functionality very clearly,
but is vulnerable to mis-clicks, may prioritize an uncommonly-used
feature, and isn't consistent with how content is typically added.

Instead, stick this in the hamburger menu, in the section that has
other page-related actions

BUG: 494649
FIXED-IN: 6.3.0