faces/applications: Include "Session Services" and "Background Services" entries
This adds the "session.slice" and "background.slice" cgroups as
(aggregated) rows called "session services" and "background services".
The rows contain all processes in either the session.slice cgroups or
the background.slice cgroups. This should result in the applications
page being a more comprehensive overview of what's actually running in a
session.