faces/applications: Filter out applications with 0 processes
If a cgroup exists but has no processes, we will list it as an
application but we won't have any relevant data to show. So just filter
these out as we can't do anything with them and the details view won't
show the correct thing either.

Ideally this should maybe be done at the ApplicationDataModel level but
that's a lot more tricky than it looks.


(cherry picked from commit 0799bb77b098868ad2ec8548247ef0c113c6d61d)

Co-authored-by: Arjen Hiemstra <ahiemstra@heimr.nl>