page: Fix re-selecting a face after closing the face configuration page
This clears the active item when the face configuration page is closed,
which allows re-opening the page by clicking the just-closed face. To
avoid breaking things when changing from one face to another, we also no
longer close and reopen the config page if a different face is selected,
instead changing the object the config page acts upon.

BUG: 500087