page: Introduce a PageManager singleton
This adds a new singleton, PageManager, to page, that contains the logic
for creating and loading pages. This makes use of the new PageController
class and replaces the code from PagesModel, because having the model
logic intertwined with the page logic makes it harder to make changes to
either code.