Move systemmonitor.kcfg to page and remove Configuration
This moves the kcfg file and code generation to Page, exposing it as a
QML type in that module. This allows the page classes to directly
interact with the configuration, instead of needing to go through some
sort of intermediary. It also exposes the configuration object as a QML
singleton, removing the need for the Configuration class.