Remove KSharedConfig from PageDataObject
There is no real need for PageDataObject to keep track of the config
object now that we have PageController. Instead give PageDataObject a
pointer to the PageController that owns it and have save() use
KConfigBase like load(). Anything requiring access to the page config
can then use PageController for that.