faces: Rework how we feed data into history line charts for tables
The row/column referenced by ComponentCacheProxyModel can change, but
the properties are marked as constant. This can lead to the wrong data
being fed into the history source, resulting in incorrect charts being
displayed.

Rather than trying to come up with some complicated method of fixing
this in ComponentCacheProxyModel, change the chart component so we can
feed it data from the actual delegate using a binding. This way we know
for certain the value put into the history matches the displayed value.