page: Fix loading of pages with empty values
Apparently, KConfig treats things differently when it loads an empty
string versus when an entry is set to an empty string. This would lead
to things not restoring correctly when a page would set some face
property to the default.

Fix this by excluding empty strings during loading, while still allowing
them to be set during saving.