Partially revert "Don't override dialog y: values"
This reverts one part of commit a0d31ca21c32c223e5e6323fc56bdfb2611c237c.

Most of these dialogs didn't need to override the y: value, but the
column configuration dialog did, since it's helpful to be able to see
which columns are currently visible.

Restore that and add a comment explaining why it's there so no one
removes it again in the future.

BUG: 505052
FIXED-IN: 6.3.6