Use unscaled CPU usage percentages by default
Scaled numbers tend to obscure problems, with the degree of obscuring
worsening as CPU core counts increase over time. For example, on a
16-core machine (common these days in consumer hardware), an app or
process pegging a core at 100% will appear to only have 6.25% CPU
usage (100% / 16 cores), obscuring the source of the issue and giving a
misleadingly rosy assessment of the system's status.

Unscaled numbers do not suffer from this issue. Since troubleshooting
problems is a primary use case of this app, let's switch to unscaled
numbers to help people out with that use case.

FEATURE: 449414
FIXED-IN: 6.4.0