Load ColumnConfigurationDialog contents on demand
The ColumnConfigurationDialog is loaded when the process table view is
opened. It contains business logic of which columns should be visible,
so moving to a Loader pattern requires a significant refactor.

Due to Kirigami Dialog this ends up loading all delegates for all rows,
not just those in view. We're loading 55 combo boxes, each with their
own drop down, which incluces a lot of invisible scrollbars on top.

Each has overhead to being around.

As an interim solution, only populate the model when visible.