page: Set smooth to false for remove margins effect
In certain cases, most notably when using fractional scaling, the
texture may end up being rendered at fractional positions. With linear
interpolation, this leads to "blurriness" of the texture. So make sure
to disable that here.

This relies on recently introduced Kirigami behaviour that makes
ShadowedTexture respect the `smooth` property.

BUG: 445759