table: reset hoveredRow if hover is gone
The hover effect was acting _weird_, because it would sometimes stay even
if the mouse has left the element. This is just a small visual glitch,
but it just didn’t look _good_, and word in the street is, KDE Plasma is
pretty, and I want it to stay that way. :smile: